refactor: [Task 1.5] Extract GameResultData and ScoreType to dedicated module
