+++ src/ui/teletext/mod.rs
@@ -1,9 +1,12 @@
 pub mod colors;
 pub mod compact_display;
+pub mod game_result;
 pub mod page_config;
 
 // Re-export for backward compatibility
 #[allow(unused_imports)]
 pub use compact_display::{CompactDisplayConfig, CompactModeValidation, TerminalWidthValidation};
+#[allow(unused_imports)]
+pub use game_result::{GameResultData, ScoreType};
 #[allow(unused_imports)]
 pub use page_config::TeletextPageConfig;
